Tweak core opcode match
We want every clause to be isolated, and to always return a result.

Ifs with a return at the end look nice, but might miss a return if cases
aren't exclusive.

If/else can do it, but with awkwardly shaped code.

Tweak the existing solution by not privileging opcode_6, have all arms
be guards. I guess that is kind of a convention too...

Best of course would be some cool bit-level pattern matching